#85 Extract RunFactory for tests
